+++ snapshots/snapshots-shared/src/main/kotlin/com/emergetools/snapshots/shared/ComposeSnapshots.kt
@@ -1,0 +1,8 @@
+package com.emergetools.snapshots.shared
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ComposeSnapshots(
+  val snapshots: List<ComposePreviewSnapshotConfig>,
+)
